+++ base/linalg/rowvector.jl
@@ -45,11 +45,6 @@ const ConjRowVector{T,CV<:ConjVector} = RowVector{T,CV}
 @inline RowVector{T}(::Uninitialized, n::Tuple{Int,Int}) where {T} =
     n[1] == 1 ? RowVector{T}(Vector{transpose_type(T)}(uninitialized, n[2])) :
         error("RowVector expects 1×N size, got $n")
-# to deprecate, RowVector{T}(shape...) constructors
-@inline RowVector{T}(n::Int) where {T} = RowVector{T}(uninitialized, n)
-@inline RowVector{T}(n1::Int, n2::Int) where {T} = RowVector{T}(uninitialized, n1, n2)
-@inline RowVector{T}(n::Tuple{Int}) where {T} = RowVector{T}(uninitializd, n)
-@inline RowVector{T}(n::Tuple{Int,Int}) where {T} = RowVector{T}(uninitialized, n)
 
 # Conversion of underlying storage
 convert(::Type{RowVector{T,V}}, rowvec::RowVector) where {T,V<:AbstractVector} =